bounce123/requirements.txt
absl-py==1.4.0
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
djangorestframework==3.14.0
idna==3.4
numpy==1.25.2
ortools==9.7.2996
protobuf==4.24.2
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
